requirements.txt
Django==5.1.7
djangorestframework==3.15.2
psycopg[binary]==3.2.6
gunicorn==23.0.0
daphne
redis==5.2.1
djangorestframework-simplejwt
django-filter
drf-spectacular
pytest
pytest-django
factory_boy
pytest-cov
aio-pika
channels
channels-redis
